*) Code from Tiv - Added: Experimental code for support of the new emblems.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/1.x@241 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/keeps/Gameobjects/GameKeepBanner.cs
@@ -132,7 +132,7 @@ namespace DOL.GS.Keeps
 		{
 			Guild guild = this.Component.Keep.Guild;
 
-			ushort emblem = 0;
+			int emblem = 0;
 			if (guild != null)
 				emblem = guild.theGuildDB.Emblem;
