+++ GameServer/keeps/Managers/Clothing Manager.cs
@@ -460,7 +460,7 @@ namespace DOL.GS.Keeps
 				return;
 			if (guard.Component == null)
 				return;
-			ushort emblem = 0;
+			int emblem = 0;
 			if (guard.Component.Keep.Guild != null) emblem = guard.Component.Keep.Guild.theGuildDB.Emblem;
 			InventoryItem cloak = guard.Inventory.GetItem(eInventorySlot.Cloak);
 			if (cloak != null) cloak.Emblem = emblem;
